% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTKalman-package.R
\docType{package}
\name{PrevKalman-package}
\alias{PrevKalman}
\alias{PrevKalman-package}
\title{PrevKalman: Horvitz-Thompson Prevalence Estimation with Kalman Smoothing}
\description{
Methods for repeated-testing prevalence estimation, delete-a-group
jackknife uncertainty quantification, and Kalman filtering/smoothing for
noisy daily prevalence trajectories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Jeongjin95/PrevKalman}
  \item Report bugs at \url{https://github.com/Jeongjin95/PrevKalman/issues}
}

}
\author{
\strong{Maintainer}: Jeongjin Lee \email{Jeongjin95@users.noreply.github.com}

}
\keyword{internal}
